scripts/requirements.txt
mysql-connector-python
pandas
feedparser
beautifulsoup4
textblob
ntscraper
google-generativeai
instaloader
facebook-scraper
